# Rebalance `CircuitInstruction` and `PackedInstruction`