--- a/qiskit/converters/circuit_to_dag.py
+++ b/qiskit/converters/circuit_to_dag.py
@@ -94,24 +94,9 @@ def circuit_to_dag(circuit, copy_operations=True, *, qubit_order=None, clbit_ord
         dagcircuit.add_creg(register)
 
     for instruction in circuit.data:
-        if not isinstance(instruction._raw_op, StandardGate):
-            op = instruction.operation
-            if copy_operations:
-                op = copy.deepcopy(op)
-            dagcircuit.apply_operation_back(op, instruction.qubits, instruction.clbits, check=False)
-        else:
-            node = DAGOpNode(
-                instruction._raw_op,
-                qargs=instruction.qubits,
-                cargs=instruction.clbits,
-                params=instruction.params,
-                label=instruction.label,
-                duration=instruction.duration,
-                unit=instruction.unit,
-                condition=instruction.condition,
-                dag=dagcircuit,
-            )
-            dagcircuit._apply_op_node_back(node)
+        dagcircuit._apply_op_node_back(
+            DAGOpNode.from_instruction(instruction, dag=dagcircuit, deepcopy=copy_operations)
+        )
 
     dagcircuit.duration = circuit.duration
     dagcircuit.unit = circuit.unit