--- a/qiskit/converters/dag_to_circuit.py
+++ b/qiskit/converters/dag_to_circuit.py
@@ -72,24 +72,7 @@ def dag_to_circuit(dag, copy_operations=True):
     circuit.calibrations = dag.calibrations
 
     for node in dag.topological_op_nodes():
-        if not isinstance(node._raw_op, StandardGate):
-            op = node.op
-            if copy_operations:
-                op = copy.deepcopy(op)
-            circuit._append(CircuitInstruction(op, node.qargs, node.cargs))
-        else:
-            circuit._append(
-                CircuitInstruction(
-                    node._raw_op,
-                    node.qargs,
-                    node.cargs,
-                    params=node.params,
-                    label=node.label,
-                    duration=node.duration,
-                    unit=node.unit,
-                    condition=node.condition,
-                )
-            )
+        circuit._append(node._to_circuit_instruction(deepcopy=copy_operations))
 
     circuit.duration = dag.duration
     circuit.unit = dag.unit