--- a/qiskit/transpiler/passes/synthesis/unitary_synthesis.py
+++ b/qiskit/transpiler/passes/synthesis/unitary_synthesis.py
@@ -42,7 +42,7 @@ from qiskit.synthesis.two_qubit.two_qubit_decompose import (
 )
 from qiskit.quantum_info import Operator
 from qiskit.circuit.controlflow import CONTROL_FLOW_OP_NAMES
-from qiskit.circuit import Gate, Parameter
+from qiskit.circuit import Gate, Parameter, CircuitInstruction
 from qiskit.circuit.library.standard_gates import (
     iSwapGate,
     CXGate,
@@ -566,17 +566,18 @@ class UnitarySynthesis(TransformationPass):
                         qargs,
                     ) in node_list:
                         if op_name == "USER_GATE":
-                            node = DAGOpNode(
-                                user_gate_node._raw_op,
-                                params=user_gate_node.params,
-                                qargs=tuple(qubits[x] for x in qargs),
+                            node = DAGOpNode.from_instruction(
+                                user_gate_node._to_circuit_instruction().replace(
+                                    params=user_gate_node.params,
+                                    qubits=tuple(qubits[x] for x in qargs),
+                                ),
                                 dag=out_dag,
                             )
                         else:
-                            node = DAGOpNode(
-                                GATE_NAME_MAP[op_name],
-                                params=params,
-                                qargs=tuple(qubits[x] for x in qargs),
+                            node = DAGOpNode.from_instruction(
+                                CircuitInstruction.from_standard(
+                                    GATE_NAME_MAP[op_name], tuple(qubits[x] for x in qargs), params
+                                ),
                                 dag=out_dag,
                             )
                         out_dag._apply_op_node_back(node)
@@ -1043,6 +1044,8 @@ class DefaultUnitarySynthesis(plugin.UnitarySynthesisPlugin):
         flip_bits = out_dag.qubits[::-1]
         for node in synth_circ.topological_op_nodes():
             qubits = tuple(flip_bits[synth_circ.find_bit(x).index] for x in node.qargs)
-            node = DAGOpNode(node._raw_op, qargs=qubits, params=node.params)
+            node = DAGOpNode.from_instruction(
+                node._to_circuit_instruction().replace(qubits=qubits, params=node.params)
+            )
             out_dag._apply_op_node_back(node)
         return out_dag