… Cargo.lock
… crates/circuit/src/operations.rs
… qiskit/dagcircuit/dagcircuit.py
… test/python/circuit/test_rust_equivalence.py