@jakelishman 2024-07-07
Benchmarking against 1.1.0, showing only changed.  Note that most utility-scale transpilations are not in this list - they're within 10% of 1.1.0 now.  `assign_parameters` is still much slower; we need to both rework `Param` handling in Rust a bit to avoid the nearly 50% padding factor, and probably move the orchestration down to Rust to avoid allocations.

```text
| Change   | Before [7d29dc1b] <1.1.0^0>   | After [d9e31ed5] <repack-instruction>   |   Ratio | Benchmark (Parameter)                                                                                           |
|----------|-------------------------------|-----------------------------------------|---------|-----------------------------------------------------------------------------------------------------------------|
| -        | 1.07±0.04ms                   | 714±90μs                                |    0.67 | circuit_construction.CircuitConstructionBench.time_circuit_construction(1, 128)                                 |
| -        | 1.04±0.01s                    | 683±30ms                                |    0.65 | circuit_construction.CircuitConstructionBench.time_circuit_construction(1, 131072)                              |
| -        | 17.1±0.8ms                    | 11.2±0.4ms                              |    0.65 | circuit_construction.CircuitConstructionBench.time_circuit_construction(1, 2048)                                |
| -        | 258±5ms                       | 172±6ms                                 |    0.67 | circuit_construction.CircuitConstructionBench.time_circuit_construction(1, 32768)                               |
| -        | 65.6±2ms                      | 42.3±2ms                                |    0.64 | circuit_construction.CircuitConstructionBench.time_circuit_construction(1, 8192)                                |
| -        | 1.22±0.01ms                   | 827±30μs                                |    0.68 | circuit_construction.CircuitConstructionBench.time_circuit_construction(14, 128)                                |
| -        | 1.15±0.01s                    | 713±20ms                                |    0.62 | circuit_construction.CircuitConstructionBench.time_circuit_construction(14, 131072)                             |
| -        | 17.6±0.3ms                    | 10.6±0.09ms                             |    0.6  | circuit_construction.CircuitConstructionBench.time_circuit_construction(14, 2048)                               |
| -        | 284±6ms                       | 183±4ms                                 |    0.64 | circuit_construction.CircuitConstructionBench.time_circuit_construction(14, 32768)                              |
| -        | 297±4μs                       | 210±10μs                                |    0.71 | circuit_construction.CircuitConstructionBench.time_circuit_construction(14, 8)                                  |
| -        | 70.1±6ms                      | 43.1±0.8ms                              |    0.61 | circuit_construction.CircuitConstructionBench.time_circuit_construction(14, 8192)                               |
| -        | 1.18±0.06ms                   | 720±10μs                                |    0.61 | circuit_construction.CircuitConstructionBench.time_circuit_construction(2, 128)                                 |
| -        | 1.11±0.01s                    | 701±5ms                                 |    0.63 | circuit_construction.CircuitConstructionBench.time_circuit_construction(2, 131072)                              |
| -        | 17.1±0.3ms                    | 10.5±0.2ms                              |    0.61 | circuit_construction.CircuitConstructionBench.time_circuit_construction(2, 2048)                                |
| -        | 277±4ms                       | 171±2ms                                 |    0.62 | circuit_construction.CircuitConstructionBench.time_circuit_construction(2, 32768)                               |
| -        | 114±0.8μs                     | 78.1±2μs                                |    0.69 | circuit_construction.CircuitConstructionBench.time_circuit_construction(2, 8)                                   |
| -        | 76.3±4ms                      | 44.4±1ms                                |    0.58 | circuit_construction.CircuitConstructionBench.time_circuit_construction(2, 8192)                                |
| -        | 1.41±0.07ms                   | 929±70μs                                |    0.66 | circuit_construction.CircuitConstructionBench.time_circuit_construction(20, 128)                                |
| -        | 1.11±0.01s                    | 702±10ms                                |    0.63 | circuit_construction.CircuitConstructionBench.time_circuit_construction(20, 131072)                             |
| -        | 17.6±0.2ms                    | 10.9±0.2ms                              |    0.62 | circuit_construction.CircuitConstructionBench.time_circuit_construction(20, 2048)                               |
| -        | 282±10ms                      | 172±3ms                                 |    0.61 | circuit_construction.CircuitConstructionBench.time_circuit_construction(20, 32768)                              |
| -        | 415±30μs                      | 296±8μs                                 |    0.71 | circuit_construction.CircuitConstructionBench.time_circuit_construction(20, 8)                                  |
| -        | 69.5±1ms                      | 42.7±0.5ms                              |    0.61 | circuit_construction.CircuitConstructionBench.time_circuit_construction(20, 8192)                               |
| -        | 1.26±0.06ms                   | 792±100μs                               |    0.63 | circuit_construction.CircuitConstructionBench.time_circuit_construction(5, 128)                                 |
| -        | 1.11±0.02s                    | 693±10ms                                |    0.63 | circuit_construction.CircuitConstructionBench.time_circuit_construction(5, 131072)                              |
| -        | 17.9±0.8ms                    | 10.8±0.2ms                              |    0.6  | circuit_construction.CircuitConstructionBench.time_circuit_construction(5, 2048)                                |
| -        | 279±4ms                       | 170±3ms                                 |    0.61 | circuit_construction.CircuitConstructionBench.time_circuit_construction(5, 32768)                               |
| -        | 123±2μs                       | 84.0±4μs                                |    0.69 | circuit_construction.CircuitConstructionBench.time_circuit_construction(5, 8)                                   |
| -        | 70.5±3ms                      | 43.9±2ms                                |    0.62 | circuit_construction.CircuitConstructionBench.time_circuit_construction(5, 8192)                                |
| -        | 1.30±0.03ms                   | 756±6μs                                 |    0.58 | circuit_construction.CircuitConstructionBench.time_circuit_construction(8, 128)                                 |
| -        | 1.14±0.04s                    | 704±5ms                                 |    0.61 | circuit_construction.CircuitConstructionBench.time_circuit_construction(8, 131072)                              |
| -        | 17.4±0.2ms                    | 11.8±0.5ms                              |    0.68 | circuit_construction.CircuitConstructionBench.time_circuit_construction(8, 2048)                                |
| -        | 282±10ms                      | 181±7ms                                 |    0.64 | circuit_construction.CircuitConstructionBench.time_circuit_construction(8, 32768)                               |
| -        | 192±8μs                       | 124±3μs                                 |    0.64 | circuit_construction.CircuitConstructionBench.time_circuit_construction(8, 8)                                   |
| -        | 71.7±3ms                      | 43.4±0.4ms                              |    0.61 | circuit_construction.CircuitConstructionBench.time_circuit_construction(8, 8192)                                |
| -        | 27.9±0.2μs                    | 10.7±0.3μs                              |    0.38 | circuit_construction.CircuitConstructionBench.time_circuit_copy(1, 128)                                         |
| -        | 18.0±0.4ms                    | 2.92±0.6ms                              |    0.16 | circuit_construction.CircuitConstructionBench.time_circuit_copy(1, 131072)                                      |
| -        | 299±9μs                       | 37.8±5μs                                |    0.13 | circuit_construction.CircuitConstructionBench.time_circuit_copy(1, 2048)                                        |
| -        | 4.59±0.3ms                    | 543±200μs                               |    0.12 | circuit_construction.CircuitConstructionBench.time_circuit_copy(1, 32768)                                       |
| -        | 12.1±0.1μs                    | 9.42±0.07μs                             |    0.78 | circuit_construction.CircuitConstructionBench.time_circuit_copy(1, 8)                                           |
| -        | 1.14±0.01ms                   | 106±20μs                                |    0.09 | circuit_construction.CircuitConstructionBench.time_circuit_copy(1, 8192)                                        |
| -        | 34.9±0.7μs                    | 16.7±0.8μs                              |    0.48 | circuit_construction.CircuitConstructionBench.time_circuit_copy(14, 128)                                        |
| -        | 17.8±0.1ms                    | 3.68±0.1ms                              |    0.21 | circuit_construction.CircuitConstructionBench.time_circuit_copy(14, 131072)                                     |
| -        | 295±2μs                       | 40.8±5μs                                |    0.14 | circuit_construction.CircuitConstructionBench.time_circuit_copy(14, 2048)                                       |
| -        | 4.56±0.2ms                    | 532±100μs                               |    0.12 | circuit_construction.CircuitConstructionBench.time_circuit_copy(14, 32768)                                      |
| -        | 20.1±0.6μs                    | 14.8±0.9μs                              |    0.73 | circuit_construction.CircuitConstructionBench.time_circuit_copy(14, 8)                                          |
| -        | 1.14±0.04ms                   | 122±30μs                                |    0.11 | circuit_construction.CircuitConstructionBench.time_circuit_copy(14, 8192)                                       |
| -        | 29.6±0.3μs                    | 12.4±0.1μs                              |    0.42 | circuit_construction.CircuitConstructionBench.time_circuit_copy(2, 128)                                         |
| -        | 17.9±0.2ms                    | 3.73±0.2ms                              |    0.21 | circuit_construction.CircuitConstructionBench.time_circuit_copy(2, 131072)                                      |
| -        | 286±3μs                       | 32.2±5μs                                |    0.11 | circuit_construction.CircuitConstructionBench.time_circuit_copy(2, 2048)                                        |
| -        | 4.61±0.2ms                    | 541±100μs                               |    0.12 | circuit_construction.CircuitConstructionBench.time_circuit_copy(2, 32768)                                       |
| -        | 12.6±0.1μs                    | 9.87±0.3μs                              |    0.78 | circuit_construction.CircuitConstructionBench.time_circuit_copy(2, 8)                                           |
| -        | 1.18±0.05ms                   | 142±20μs                                |    0.12 | circuit_construction.CircuitConstructionBench.time_circuit_copy(2, 8192)                                        |
| -        | 41.0±0.6μs                    | 19.5±1μs                                |    0.48 | circuit_construction.CircuitConstructionBench.time_circuit_copy(20, 128)                                        |
| -        | 18.2±0.3ms                    | 3.67±0.1ms                              |    0.2  | circuit_construction.CircuitConstructionBench.time_circuit_copy(20, 131072)                                     |
| -        | 312±20μs                      | 50.9±6μs                                |    0.16 | circuit_construction.CircuitConstructionBench.time_circuit_copy(20, 2048)                                       |
| -        | 4.62±0.1ms                    | 577±100μs                               |    0.12 | circuit_construction.CircuitConstructionBench.time_circuit_copy(20, 32768)                                      |
| -        | 1.15±0.02ms                   | 117±8μs                                 |    0.1  | circuit_construction.CircuitConstructionBench.time_circuit_copy(20, 8192)                                       |
| -        | 30.0±0.3μs                    | 12.9±0.2μs                              |    0.43 | circuit_construction.CircuitConstructionBench.time_circuit_copy(5, 128)                                         |
| -        | 18.1±0.1ms                    | 3.62±0.3ms                              |    0.2  | circuit_construction.CircuitConstructionBench.time_circuit_copy(5, 131072)                                      |
| -        | 309±6μs                       | 34.0±3μs                                |    0.11 | circuit_construction.CircuitConstructionBench.time_circuit_copy(5, 2048)                                        |
| -        | 4.63±0.2ms                    | 576±100μs                               |    0.12 | circuit_construction.CircuitConstructionBench.time_circuit_copy(5, 32768)                                       |
| -        | 14.9±0.2μs                    | 10.5±0.1μs                              |    0.71 | circuit_construction.CircuitConstructionBench.time_circuit_copy(5, 8)                                           |
| -        | 1.19±0.03ms                   | 108±30μs                                |    0.09 | circuit_construction.CircuitConstructionBench.time_circuit_copy(5, 8192)                                        |
| -        | 32.0±0.8μs                    | 14.9±2μs                                |    0.47 | circuit_construction.CircuitConstructionBench.time_circuit_copy(8, 128)                                         |
| -        | 18.1±0.6ms                    | 3.61±0.3ms                              |    0.2  | circuit_construction.CircuitConstructionBench.time_circuit_copy(8, 131072)                                      |
| -        | 288±5μs                       | 44.5±8μs                                |    0.15 | circuit_construction.CircuitConstructionBench.time_circuit_copy(8, 2048)                                        |
| -        | 4.42±0.1ms                    | 484±200μs                               |    0.11 | circuit_construction.CircuitConstructionBench.time_circuit_copy(8, 32768)                                       |
| -        | 1.10±0.02ms                   | 118±20μs                                |    0.11 | circuit_construction.CircuitConstructionBench.time_circuit_copy(8, 8192)                                        |
| -        | 187±5μs                       | 30.9±0.3μs                              |    0.17 | circuit_construction.CircuitConstructionBench.time_circuit_extend(1, 128)                                       |
| -        | 168±2ms                       | 23.0±0.9ms                              |    0.14 | circuit_construction.CircuitConstructionBench.time_circuit_extend(1, 131072)                                    |
| -        | 2.75±0.1ms                    | 343±8μs                                 |    0.12 | circuit_construction.CircuitConstructionBench.time_circuit_extend(1, 2048)                                      |
| -        | 41.7±0.3ms                    | 5.68±0.2ms                              |    0.14 | circuit_construction.CircuitConstructionBench.time_circuit_extend(1, 32768)                                     |
| -        | 20.3±0.5μs                    | 11.3±0.5μs                              |    0.56 | circuit_construction.CircuitConstructionBench.time_circuit_extend(1, 8)                                         |
| -        | 11.1±0.3ms                    | 1.39±0.08ms                             |    0.13 | circuit_construction.CircuitConstructionBench.time_circuit_extend(1, 8192)                                      |
| -        | 204±2μs                       | 54.1±7μs                                |    0.26 | circuit_construction.CircuitConstructionBench.time_circuit_extend(14, 128)                                      |
| -        | 184±2ms                       | 34.7±0.2ms                              |    0.19 | circuit_construction.CircuitConstructionBench.time_circuit_extend(14, 131072)                                   |
| -        | 2.87±0.1ms                    | 439±10μs                                |    0.15 | circuit_construction.CircuitConstructionBench.time_circuit_extend(14, 2048)                                     |
| -        | 44.9±0.5ms                    | 7.24±0.4ms                              |    0.16 | circuit_construction.CircuitConstructionBench.time_circuit_extend(14, 32768)                                    |
| -        | 53.2±0.6μs                    | 27.0±0.7μs                              |    0.51 | circuit_construction.CircuitConstructionBench.time_circuit_extend(14, 8)                                        |
| -        | 11.4±0.3ms                    | 1.83±0.09ms                             |    0.16 | circuit_construction.CircuitConstructionBench.time_circuit_extend(14, 8192)                                     |
| -        | 188±2μs                       | 36.0±2μs                                |    0.19 | circuit_construction.CircuitConstructionBench.time_circuit_extend(2, 128)                                       |
| -        | 187±6ms                       | 33.6±2ms                                |    0.18 | circuit_construction.CircuitConstructionBench.time_circuit_extend(2, 131072)                                    |
| -        | 2.98±0.09ms                   | 380±20μs                                |    0.13 | circuit_construction.CircuitConstructionBench.time_circuit_extend(2, 2048)                                      |
| -        | 45.1±1ms                      | 7.02±0.3ms                              |    0.16 | circuit_construction.CircuitConstructionBench.time_circuit_extend(2, 32768)                                     |
| -        | 23.3±1μs                      | 11.5±0.2μs                              |    0.49 | circuit_construction.CircuitConstructionBench.time_circuit_extend(2, 8)                                         |
| -        | 11.9±0.3ms                    | 1.62±0.05ms                             |    0.14 | circuit_construction.CircuitConstructionBench.time_circuit_extend(2, 8192)                                      |
| -        | 246±8μs                       | 57.8±10μs                               |    0.24 | circuit_construction.CircuitConstructionBench.time_circuit_extend(20, 128)                                      |
| -        | 182±1ms                       | 35.6±2ms                                |    0.2  | circuit_construction.CircuitConstructionBench.time_circuit_extend(20, 131072)                                   |
| -        | 2.91±0.05ms                   | 447±7μs                                 |    0.15 | circuit_construction.CircuitConstructionBench.time_circuit_extend(20, 2048)                                     |
…